Add backend_servers pillar, for static backend defn
The existing backend definition requires enabling Salt Mine,
and a somewhat laborious grain configuration. This is good if a
dynamic load balancer is required, but overkill if the LB just maintains
a static list and adds/drops nodes based on health check results.

This commit allows us to alternatively use a static configuration (or
indeed, a bit of both).