Allow protection of backend via basic auth acl
This is a simple implementation of basic auth on the haproxy balancer
to allow us to block sites entirely with a username/password. This is
commonly done on preview/staging environments to protect from accidental
use by the public, without the need for firewall restrictions.

The default is for it to be 'off', turn on as needed.

This has been implemented as per: <http://serverfault.com/questions/239749/possible-to-add-basic-http-access-authentication-via-haproxy>

... which in turn references <https://nbevans.wordpress.com/2011/03/03/cultural-learnings-of-ha-proxy-for-make-benefit>